STEP3/SQL_PYTHON/requirements.txt
Faker==15.3.3
mysql==0.0.3
mysql-client==0.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
paho-mqtt==1.6.1
protobuf==3.20.1
python-dateutil==2.8.2
six==1.16.0
